requirements.txt
asgiref==3.4.1
cefpython3==66.1
Django==4.0.1
django-solo==2.0.0
Pillow==9.0.0
sqlparse==0.4.2
tzdata==2021.5
